+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Version 1.0.0-a1 (Initial Version)
 
 - Support the following commands of the API v3.0:
+   - `VLCCommandNameNext`: Jumps to next playlist item.
    - `VLCCommandNamePlay`: Plays a specific playlist item.
    - `VLCCommandNameSeek`: Seek to a particular point in the current playback
    - `VLCCommandNameStatus`: Retrieve the player status informations